requirements.txt
Django==3.0.7
djangorestframework==3.11.0
schedule==0.6.0
